fix modif pba et debug fus

+++ code/AI/ai_lists/analyse/calc_proba_trivial/calc_proba_trivial.h
@@ -3,7 +3,7 @@
 
 #include "../../../../fundamental_functions/game_functions/access_functions/access_functions.h"
 
-float esperanceLignesPions(Game *g, bool color);
-float varianceLignesPions(Game *g, bool color);
+double esperanceLignesPions(Game *g, bool color);
+double varianceLignesPions(Game *g, bool color);
 
 #endif
